Explicit casting

+++ ptp_comms/src/serialisers.cpp
@@ -41,32 +41,32 @@ namespace serialisers
     uint16_t getU16(uint8_t* src)
     {
         uint16_t val = 0;
-        val |= (src[0] & 0xff) << 8;
-        val |= (src[1] & 0xff);
+        val |= (uint16_t)(src[0] & 0xff) << 8;
+        val |= (uint16_t)(src[1] & 0xff);
         return val;
     }
 
     uint32_t getU32(uint8_t* src)
     {
         uint32_t val = 0;
-        val |= (src[0] & 0xff) << 24;
-        val |= (src[1] & 0xff) << 16;
-        val |= (src[2] & 0xff) << 8;
-        val |= (src[3] & 0xff);
+        val |= (uint32_t)(src[0] & 0xff) << 24;
+        val |= (uint32_t)(src[1] & 0xff) << 16;
+        val |= (uint32_t)(src[2] & 0xff) << 8;
+        val |= (uint32_t)(src[3] & 0xff);
         return val;
     }
 
     uint64_t getU64(uint8_t* src)
     {
         uint64_t val = 0;
-        val |= (src[0] & 0xff) << 56;
-        val |= (src[1] & 0xff) << 48;
-        val |= (src[2] & 0xff) << 40;
-        val |= (src[3] & 0xff) << 32;
-        val |= (src[4] & 0xff) << 24;
-        val |= (src[5] & 0xff) << 16;
-        val |= (src[6] & 0xff) << 8;
-        val |= (src[7] & 0xff);
+        val |= (uint64_t)(src[0] & 0xff) << 56;
+        val |= (uint64_t)(src[1] & 0xff) << 48;
+        val |= (uint64_t)(src[2] & 0xff) << 40;
+        val |= (uint64_t)(src[3] & 0xff) << 32;
+        val |= (uint64_t)(src[4] & 0xff) << 24;
+        val |= (uint64_t)(src[5] & 0xff) << 16;
+        val |= (uint64_t)(src[6] & 0xff) << 8;
+        val |= (uint64_t)(src[7] & 0xff);
         return val;
     }
 }
